ssj-load-balancer/requirements.txt
tqdm
numpy~=1.19.1
kafka-python==2.0.2
requests~=2.26.0
minio~=7.1.1